repo cleaning and minor fixes
