+++ .travis.yml
@@ -1,0 +1,35 @@
+matrix:
+  include:
+    - language:  php
+      php: 5.6
+      before_script:
+          - find . -type f -name *.php | xargs -n1 php -l
+      script:        
+          - cd ${TRAVIS_BUILD_DIR}
+          - pwd
+      after_success:
+          - cd ${TRAVIS_BUILD_DIR}
+          - ls -latr
+      after_failure:
+          - cd ${TRAVIS_BUILD_DIR}
+          - ls -latr
+    
+    - language: python
+      python: 2.7
+      install:
+        - pip install pylint
+      script:
+        - cd ${TRAVIS_BUILD_DIR}
+        - ./tests/tools/lintAllPythonFiles.sh
+    - language: markdown
+      addons:
+        apt:
+          packages:
+            - aspell
+            - aspell-fr
+      script:
+        - gem install mdl
+        - cd ${TRAVIS_BUILD_DIR}
+        - ./tests/tools/setCustomMDWarnings.sh 
+        - mdl -r $MDLWAR *.md docs/fr_FR/*.md
+        - ./tests/tools/spellCheckMD.sh
