+++ ISSUE_TEMPLATE.md
@@ -1,0 +1,18 @@
+## Description
+
+## Etapes à reproduire (pour les bugs)
+
+1.
+2.
+3.
+4.
+
+## Contexte:
+
+## Proposition de solution (optionnel):
+
+## Environnement:
+
+* **Version Jeedom**:
+* **Plateforme**:
+* **Version du Plugin**:
