+++ resources/install_check.sh
@@ -1,4 +1,3 @@
-rm -f /tmp/dependancycheck_googlecast
 if ! python3 -V 2>&1 | grep -q "Python 3"; then
     echo "nok"
     exit 0
@@ -8,12 +7,12 @@ pip3cmd=$(compgen -ac | grep -E '^pip-?3' | sort -r | head -1)
 if [[ !  -z  $pip3cmd  ]]; then     # pip3 found
     $(sudo $pip3cmd list 2>/dev/null | grep -E "zeroconf|requests|protobuf" | wc -l > /tmp/dependancycheck_googlecast)
     content=$(cat /tmp/dependancycheck_googlecast 2>/dev/null)
+    rm -f /tmp/dependancycheck_googlecast
     if [[ -z  $content  ]]; then
-        $content = 0
+        content=0
     fi
     if [ "$content" -lt 3  ];then
         echo "nok"
-        rm -f /tmp/dependancycheck_googlecast
         exit 0
     fi
 else
@@ -21,5 +20,4 @@ else
     exit 0
 fi
 echo "ok"
-rm -f /tmp/dependancycheck_googlecast
 exit 0
