+++ tests/tools/.aspell.fr.pws
@@ -1,0 +1,25 @@
+personal_ws-1.1 fr 0 utf-8
+Etapes
+plateforme
+docs
+fr
+FR
+presentation.md
+https
+téléchargement
+prévisualisation
+xxxx
+FAQ
+releases
+release
+bold
+market
+desktop
+bug
+bugtracker
+bugs
+changelog
+config
+jeedom
+plugin
+screenshot
