+++ tests/tools/.pylintrc
@@ -1,0 +1,4 @@
+[BASIC]
+
+const-rgx=[a-z_][a-z0-9_]{2,30}$
+module-rgx=(([a-z_][a-z0-9_]*)|([A-Z][a-zA-Z0-9]+))$
