+++ tests/tools/setCustomMDWarnings.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+export MDLWAR="MD003,MD004,MD005,MD006,MD007,MD008,MD009,MD010,MD011,MD012,MD014,MD018,MD019,MD020,MD021,MD022,MD023,MD024,MD027,MD028,MD029,MD030,MD031,MD032,MD033,MD034,MD035,MD036,MD038,MD039"
